move eligibility constants around
